pkg/download: for downloading files into tmp
cmd/podman needs that functionality as well, so I figured it makes sense
to break it into a separate package to reduce code clones.
